\name{kde}
\alias{kde}
\title{
  Univariate kernel density estimation directly in R code.
}
\description{
  This function behaves similarly to the \code{density} function of the
  \pkg{stats} package, but uses only R code.  It is a demonstration
  function intended to show how kernel density estimates are
  computed, at least conceptually.  Unlike \code{density}, the kernel
  may be supplied as an R function in a standard form.  Example kernel
  functions are provided.  For computational efficiency, the
  \code{density} function of the \pkg{stats} package is far superior.
}
\usage{
kde(x, bw = bw.nrd0, kernel = kernelGaussian, n = 512,
    from = min(x) - cut * sd, to = max(x) + cut * sd,
    adjust = 1, cut = 3, ...)
}
\arguments{
  \item{x}{
    Univeriate sample.  Must be numeric.
  }
  \item{bw}{
    Either an explicit numeric bandwidth to be used for the kernel, or a
    function used to calculate it.
  }
  \item{kernel}{
    The kernel function to be used.  Must have the same argument
    sequence as \code{\link{kernelGaussian}}, with the same meanings.
  }
  \item{n}{
    Then number of points covering the range at which to evaluate the
    KDE.  More gives a smoother display of the result; fewer gives a
    quicker and more memory efficient computation.
  }
  \item{from}{
    Lower boundary for the computed KDE.
  }
  \item{to}{
    Upper boundary for the computed KDE.
  }
  \item{adjust}{
    Adjustment factor to be used for the bandwidth.
  }
  \item{cut}{
    Number of bandwidths by which to extend the range of
    the data for the range of the KDE
  }
  \item{\dots}{
    Additional arguments, if needed, to be supplied to the kernel
    function.
  }
}
\details{
  This is a demonstration function intended to show, via R code, the way
  in which a kernel density estimate is computed.

  For samples which are not too large the computation is reasonably
  efficient, but for serious computations the standard function
  \code{\link[stats]{density}}, or some alternative, should be
  used.
}
\value{
  An object of class \dQuote{density}, with essentially the same
  structure as objects generated by the \code{density}
  of the \pkg{stats} package.  \code{plot} and allied methods should apply.
}
\author{
  Bill Venables
}
\note{
  Demonstration code only!
}

\seealso{
  \code{\link{kernelBiweight}} and aliases; \code{\link[stats]{density}}.
}
\examples{
if(require("graphics")) {
  with(MASS::geyser, {
      hist(waiting, freq=FALSE, main="", border="grey", las=1)
      lines(stats::density(waiting), col="skyblue", lwd=8)
      lines(kde(waiting))
      lines(kde(waiting, kernel = kernelUniform), col="red")
      rug(jitter(waiting), col="blue")
      legend("topleft", c("density histogram",
        "KDE gaussian (denstiy)", "KDE gaussian (kde)",
        "KDE rectangular (kde)"), lty = "solid", lwd=c(1,8,1,1),
        col=c("grey", "skyblue", "black", "red"), bty="n")
  })
}
}
\keyword{distribution}
\keyword{dplot}
